+++ src/main/java/me/xhsun/guildwars2wrapper/requests/v2/backstory/Backstory.java
@@ -1,18 +1,20 @@
 package me.xhsun.guildwars2wrapper.requests.v2.backstory;
 
-import java.util.List;
 import me.xhsun.guildwars2wrapper.GuildWars2;
 import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
 import me.xhsun.guildwars2wrapper.model.v2.backstory.BackStoryAnswer;
+import me.xhsun.guildwars2wrapper.model.v2.backstory.BackStoryQuestion;
 import me.xhsun.guildwars2wrapper.requests.GuildWars2API;
 import me.xhsun.guildwars2wrapper.requests.Request;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class Answers extends Request {
+import java.util.List;
 
-	protected Answers(GuildWars2API gw2API) {
+public class Backstory extends Request {
+
+	protected Backstory(GuildWars2API gw2API) {
 		super(gw2API);
 	}
 
@@ -42,6 +44,33 @@ public class Answers extends Request {
 	public void getBackStoryAnswerInfo(String[] ids, Callback<List<BackStoryAnswer>> callback) throws GuildWars2Exception, NullPointerException {
 		isParamValid(new ParamChecker(ids));
 		gw2API.getBackStoryAnswerInfo(processIds(ids), GuildWars2.lang.getValue()).enqueue(callback);
+	}
+
+	/**
+	 * For more info on back story questions API go <a href="https://wiki.guildwars2.com/wiki/API:2/backstory/questions">here</a><br/>
+	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
+	 *
+	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
+	 * @throws NullPointerException if given {@link Callback} is empty
+	 * @see BackStoryQuestion back story question info
+	 */
+	public void getAllBackStoryQuestionID(Callback<List<Integer>> callback) throws NullPointerException {
+		gw2API.getAllBackStoryQuestionIDs().enqueue(callback);
+	}
+
+	/**
+	 * For more info on back story questions API go <a href="https://wiki.guildwars2.com/wiki/API:2/backstory/questions">here</a><br/>
+	 * Give user the access to {@link Callback#onResponse(Call, Response)} and {@link Callback#onFailure(Call, Throwable)} methods for custom interactions
+	 *
+	 * @param ids      list of back story question id(s)
+	 * @param callback callback that is going to be used for {@link Call#enqueue(Callback)}
+	 * @throws GuildWars2Exception  invalid API key
+	 * @throws NullPointerException if given {@link Callback} is empty
+	 * @see BackStoryQuestion back story question info
+	 */
+	public void getBackStoryQuestionInfo(int[] ids, Callback<List<BackStoryQuestion>> callback) throws GuildWars2Exception, NullPointerException {
+		isParamValid(new ParamChecker(ids));
+		gw2API.getBackStoryQuestionInfo(processIds(ids), GuildWars2.lang.getValue()).enqueue(callback);
 	}
 
 	//SYNC
